build: update Gradle wrapper version
gradle-wrapper.properties file has been updated to reference a new Gradle version. This change was for updating the project to use version 8.2.1 of Gradle instead of version 8.1.1. An additional line for URL validation has also been added to improve security by ensuring that the download link for the Gradle distribution is always valid.